Reorganize Product and ProductReview models into ecommerce/product/ subfolder
