+++ src/main/java/es/marcha/backend/controller/ecommerce/ProductController.java
@@ -18,9 +18,9 @@ import es.marcha.backend.dto.request.ecommerce.ProductRequestDTO;
 import es.marcha.backend.dto.response.ecommerce.ProductResponseDTO;
 import es.marcha.backend.dto.response.ecommerce.ProductReviewResponseDTO;
 import es.marcha.backend.mapper.ProductMapper;
-import es.marcha.backend.model.ecommerce.Product;
-import es.marcha.backend.model.ecommerce.ProductReview;
 import es.marcha.backend.model.ecommerce.Subcategory;
+import es.marcha.backend.model.ecommerce.product.Product;
+import es.marcha.backend.model.ecommerce.product.ProductReview;
 import es.marcha.backend.services.ecommerce.ProductReviewService;
 import es.marcha.backend.services.ecommerce.ProductService;
 import es.marcha.backend.services.ecommerce.SubcategoryService;
