+++ src/main/java/es/marcha/backend/dto/response/order/OrderItemsResponseDTO.java
@@ -2,7 +2,7 @@ package es.marcha.backend.dto.response.order;
 
 import java.math.BigDecimal;
 
-import es.marcha.backend.model.ecommerce.Product;
+import es.marcha.backend.model.ecommerce.product.Product;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
