+++ src/main/java/es/marcha/backend/mapper/ProductReviewMapper.java
@@ -3,7 +3,7 @@ package es.marcha.backend.mapper;
 import es.marcha.backend.dto.response.ecommerce.ProductReviewResponseDTO;
 import es.marcha.backend.dto.response.ecommerce.ProductReviewResponseProductDTO;
 import es.marcha.backend.dto.response.ecommerce.ProductReviewUserResponseDTO;
-import es.marcha.backend.model.ecommerce.ProductReview;
+import es.marcha.backend.model.ecommerce.product.ProductReview;
 import es.marcha.backend.model.user.User;
 
 public class ProductReviewMapper {
