+++ src/main/java/es/marcha/backend/model/ecommerce/Inventory.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
+import es.marcha.backend.model.ecommerce.product.Product;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
