+++ src/main/java/es/marcha/backend/model/ecommerce/product/ProductReview.java
@@ -1,4 +1,4 @@
-package es.marcha.backend.model.ecommerce;
+package es.marcha.backend.model.ecommerce.product;
 
 import java.time.LocalDateTime;
