+++ src/main/java/es/marcha/backend/repository/ecommerce/ProductRepository.java
@@ -3,7 +3,7 @@ package es.marcha.backend.repository.ecommerce;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import es.marcha.backend.model.ecommerce.Product;
+import es.marcha.backend.model.ecommerce.product.Product;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
