+++ src/main/java/es/marcha/backend/repository/ecommerce/ProductReviewRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import es.marcha.backend.model.ecommerce.ProductReview;
+import es.marcha.backend.model.ecommerce.product.ProductReview;
 
 @Repository
 public interface ProductReviewRepository extends JpaRepository<ProductReview, Long> {
