+++ src/main/java/es/marcha/backend/services/ecommerce/ProductReviewService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 import es.marcha.backend.dto.response.ecommerce.ProductReviewResponseDTO;
 import es.marcha.backend.exception.ProductException;
 import es.marcha.backend.mapper.ProductReviewMapper;
-import es.marcha.backend.model.ecommerce.ProductReview;
+import es.marcha.backend.model.ecommerce.product.ProductReview;
 import es.marcha.backend.model.user.User;
 import es.marcha.backend.repository.ecommerce.ProductReviewRepository;
 import es.marcha.backend.services.user.UserService;
